Created UAppRecord class, added obs-style UAppID calc, added JavaDocs.
I've ported the calculation of a UAppID string from the other work in-lab. The old code was using just a single certificate hash and a separate hash of the APK. It referred to the APK hash as the UAppID but for reasons that are unclear to me. I suspect the binary APK hash should be integrated into the calculation of the UAppID.

In order to contain the package name, the APK hash, the APK filepath and the UAppID in one bundle I created a POJO UAppRecord bean. Its just a dummy data store, but could eventually be shoved into a SQLite database perhaps.

I generated & included the built javadoc documentation. It's not perfect, but its something! I figured checking it in will make it easier to hyperlink to.